requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
pydantic==2.6.0
openai>=1.0.0
numpy>=1.21.0
python-dotenv>=1.0.0
pyyaml>=6.0
